From pgn-read-1.2.4.tar.gz uploaded to solentware.co.uk on 2019-05-21.
Fix the validation of non-pawn moves or captures something so the move or
capture is spotted as illegal if the specified start file or rank does not
match the origin square.  For example the move Rfe3 is now illegal where
before it was treated as R1e3 with rooks on e1 and f2 and squares e2 and
e3 empty.