From pgn-read-1.3.1.tar.gz uploaded to solentware.co.uk on 2019-11-14.
Games containing valid PGN movetext for moves such as Qb2c3 or Qb2xc3 no
longer cause a program crash when processed by the get_export_* methods
in PGNDisplay class.

The move counter, for half-moves since last capture or pawn move, and
full-move counter, for a white move and a black move, are added to the
data describing each position in a game.  (This is added now, rather than
ages ago, as the easiest way to enable ChessTab to display and navigate
the longest possible game in reasonable time.)