From pgn-read-2.0.tar.gz uploaded to solentware.co.uk on 2020-09-02.
Rewrite to support strict interpretation of PGN and a number of
relaxations of strictness.

The PGN standard prohibits unnecessary precision in movetext: for example
'Nge2' where 'Ne2' is sufficient.  This prohibition can be ignored, going
as far as allowing Long Algebraic Notation except for pawn moves: 'c2c4'
is not allowed but 'c2-c4' is allowed, and c2xb3 is allowed.  The Short
Algebraic Notation equivalents are 'c4' and 'cxb3'.

Lower and upper case characters can be allowed in movetext.  Thus 'D4'
can be a pawn move and 'rE7' can be a rook move.  Some bishop and pawn
moves may be ambiguous and case may have to matter in deciding between
the bishop and pawn move interpretations.

Sequences of characters which are not valid PGN movetext can be allowed
and would be ignored.

The reserved, '<...>', and escaped, lines starting '%', sequences of
characters can be captured rather than ignored.

The FIDE notations for castling and pawn promotion are different to PGN
and can be allowed.

Functions too closely associated with the ChessTab database are removed
(to ChessTab).