From pgn-read-2.0.2.tar.gz uploaded to solentware.co.uk on 2021-01-02.
Remove check indicators, and translate traditional annotations such as '!!'
to the Numberic Annotation Glyph equivalents, on import so these can be
added to PGN output in a consistent way.

Fix problem handling '{...}' and '<...>' sequences which contain Game
Termination Marker text or are missing the '}' or '>' character at end.

Attempt to fix PGN tag values in which '"' and '\' characters seem to be
not escaped when some relaxation of strictness is allowed.  The correctly
escaped version of "Team "C"" is 'Team \"C\""' for example.

Attempt to fit the PGN specification of default tag values more closely.
The value "" is accepted unless a default value is given.  Thus only the
Seven Tag Roster tags and the 'WhiteTitle', 'BlackTitle', 'WhiteElo',
'BlackElo', 'WhiteNA', and 'BlackNA', tags are required to have a non-""
value in a valid PGN game.