Fix bishop and b-pawn processing when ignoring case.
Treat 'B' not followed by a rank indicator, and 'b' followed by a file
indicator, as a bishop move.  Treat other cases as a pawn move.

The 'c4c27522..' commit for this redefined ignore case; here the rules for
'B' and 'b' are adjusted.

The FIDE notation for pawn promotion is not supported when ignoring case.