Fix disambiguate captures when ignoring case.
Forgot to change x to [xX] for disambiguate clause in ignore case regular
expression.

The upper case tests on Little.pgn changed their token counts, but not the
reported an error if there is one.

The disambiguate, RAV, and FEN, tests are renumbered so test_nnn(_mm)_*
method names are unique except where overridden.