Try [Bb][1-8] as pawn move and 'bishop on rank' when ignoring case.
[Bb][1-8] is assumed to be a pawn move at first when ignoring case. If
not a pawn move, combine it with the next token and try it as a bishop
move.  If this fails too it is an error.

Many tests on Little.pgn changed their token counts and the reported token
where there is a PGN error.  However all the tests which should see a game
without errors now do so.