Fix bishop and b-pawn processing when ignoring case.
Add flexibility of expression and resolution of ambiguity, when ignoring
case, for pawn moves on, and bishop moves to, the b-file; and for b-pawn
and bishop captures on the a-file and c-file.

Add tests for the new cases.

Some existing tests for ignore case give different results: several move
the point of detection of a PGN error towards the end of the game, and the
rest now say no PGN errors detected.