Verify king is not in check after, not before, updating game state.
Undo the game state change if the king is in check.

The classes in core.game module are moved so each is in a module of it's
own.  Methods which manipulate game state are moved to a base class.
Methods which decide what manipulation is needed remain in the other
classes.  A number of methods are renamed for clarity.

A couple of cases, for long algebraic notation piece moves, still verify
king is not in check before updating game state because unittest errors
occur otherwise when undoing the game state change.

New version 2.5.dev2 of pgn-read.